Last Update - Maintenance Completed.
Ready To Use

+++ rentACar/src/main/java/com/RentACar/business/dtos/ListAdditionalServiceDto.java
@@ -12,6 +12,6 @@ public class ListAdditionalServiceDto {
 	private int serviceId;
 	private String serviceName;
 	private String serviceDescription;
-	private int servicePrice;
+	private double servicePrice;
 	
 }
